Refactor: InputSelect ausgelagert
