+++ sys-src/frontend/src/components/Signup.tsx
@@ -5,6 +5,7 @@ import axios from 'axios';
 import InputLabel from './InputLabel'
 import InputField from "./InputField";
 import InputTextarea from "./InputTextarea";
+import InputSelect from "./InputSelect";
 
 /**
  * @interface InputFieldProps Datentyp
@@ -57,41 +58,7 @@ const InputFieldWithFixedText: React.FC<InputFieldProps & {fixedText:string}> = 
 };
 
 
-interface InputSelectProps {
-    fieldWidth: number;
-    id: string;
-    label: string;
-    // register: any;
-    required: boolean;
-    selectOptions:Array<string>
-    autoComplete?: string;
-    error?: string;
-}
 
-const InputSelect: React.FC<InputSelectProps> = ({...probs}) => {
-
-    // Variablen:
-    let classNameWidth= probs.fieldWidth ? 'sm:col-span-' + probs.fieldWidth : 'col-span-full';
-
-    // HTML:
-    return (
-        <div className={classNameWidth}>
-            <InputLabel htmlFor={probs.id} label={probs.label}/>
-            <div className="mt-2">
-                <select id={probs.id} autoComplete={probs.autoComplete} required={probs.required}
-                        className="px-2 py-2 rounded-lg block w-full
-                                    text-gray-900 placeholder:text-gray-400
-                                    border-0 shadow-sm
-                                    ring-1 ring-inset ring-gray-300
-                                    focus:outline-none focus:ring-1 focus:ring-indigo-700">
-                    {probs.selectOptions.map(elem =>
-                    <option key={elem} value={elem}>{elem}</option>)
-                    }
-                </select>
-            </div>
-        </div>
-    );
-};
